+++ src/test/java/uk/gov/hmcts/reform/roleassignment/controller/endpoints/CreateAssignmentControllerTest.java
@@ -1,15 +1,44 @@
 package uk.gov.hmcts.reform.roleassignment.controller.endpoints;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+import org.springframework.http.ResponseEntity;
+import uk.gov.hmcts.reform.roleassignment.domain.model.AssignmentRequest;
+import uk.gov.hmcts.reform.roleassignment.domain.model.enums.Status;
+import uk.gov.hmcts.reform.roleassignment.domain.service.createroles.CreateRoleAssignmentOrchestrator;
+import uk.gov.hmcts.reform.roleassignment.helper.TestDataBuilder;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
 
 class CreateAssignmentControllerTest {
 
-    // This test should be removed later
+    @Mock
+    private CreateRoleAssignmentOrchestrator createRoleAssignmentServiceMock =
+        mock(CreateRoleAssignmentOrchestrator.class);
+
+    @InjectMocks
+    private CreateAssignmentController sut;
+
+    @BeforeEach
+    public void setUp() {
+        MockitoAnnotations.initMocks(this);
+    }
+
     @Test
-    void sampleTest() {
-        String caseId = "1234567812345678";
-        assertEquals("1234567812345678", caseId);
+    void createRoleAssignment() throws Exception {
+        AssignmentRequest request = TestDataBuilder.buildAssignmentRequest(Status.CREATED, Status.LIVE, false);
+        ResponseEntity<Object> expectedResponse
+            = TestDataBuilder.buildRoleAssignmentResponse(Status.CREATED, Status.LIVE, false);
+        when(createRoleAssignmentServiceMock.createRoleAssignment(request)).thenReturn(expectedResponse);
+        ResponseEntity<Object> response = sut.createRoleAssignment(request);
+        assertNotNull(response);
+        assertEquals(expectedResponse.getStatusCode(), response.getStatusCode());
+        assertEquals(expectedResponse.getBody(), response.getBody());
     }
 }
